+++ Tests/TEST_StreamAPITests/StreamAPITests.cs
@@ -108,7 +108,8 @@ namespace TEST_StreamAPITests
             bot.Message("Testing external axes");
 
             // setup
-            bot.ExternalAxesTo(1800);
+            //bot.ExternalAxesTo(1800);
+            bot.ExternalAxisTo(1, 1800);
 
             // Home
             bot.SpeedTo(500);